Deploys static export for DigitalOcean compatibility
Changes the build output from standalone to export to resolve
dependency issues with pnpm on DigitalOcean.

This commit also includes a startup script update to use the `serve`
package for serving the static files from the 'out' directory, which
is the output directory for the exported application.

This resolves deployment issues related to the previous standalone
output and simplifies the startup process on DigitalOcean.