Prevents redirect loops after sign-in
Modifies sign-in page to display a message with a link to the dashboard if the user is already authenticated.

This avoids redirect loops caused by the middleware redirecting authenticated users to the sign-in page, which then redirects them back to the dashboard.

Also, temporarily adds dashboard route to public routes in middleware to prevent redirection for already authenticated users.